# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "ContactIn": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "notify_on_sos": {
            "default": true,
            "title": "Notify On Sos",
            "type": "boolean"
          },
          "phone": {
            "maxLength": 20,
            "minLength": 7,
            "title": "Phone",
            "type": "string"
          },
          "relationship_label": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relationship Label"
          }
        },
        "required": [
          "name",
          "phone"
        ],
        "title": "ContactIn",
        "type": "object"
      },
      "ContactOut": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notify_on_sos": {
            "title": "Notify On Sos",
            "type": "boolean"
          },
          "phone": {
            "title": "Phone",
            "type": "string"
          },
          "priority": {
            "title": "Priority",
            "type": "integer"
          },
          "relationship_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relationship Label"
          }
        },
        "required": [
          "id",
          "name",
          "phone",
          "relationship_label",
          "priority",
          "notify_on_sos"
        ],
        "title": "ContactOut",
        "type": "object"
      },
...
      "EmergencyProfileModel": {
        "additionalProperties": false,
        "properties": {
          "advance_directive": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Advance Directive"
          },
          "critical_information": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Critical Information"
          },
          "organ_donor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OrganDonorStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "show_allergies": {
            "default": true,
            "title": "Show Allergies",
            "type": "boolean"
          },
          "show_blood_group": {
            "default": true,
            "title": "Show Blood Group",
            "type": "boolean"
          },
          "show_conditions": {
            "default": false,
            "title": "Show Conditions",
            "type": "boolean"
          },
          "show_medications": {
            "default": false,
            "title": "Show Medications",
            "type": "boolean"
          }
        },
        "title": "EmergencyProfileModel",
        "type": "object"
      },
      "EmergencyProfileOut": {
        "additionalProperties": false,
        "properties": {
          "advance_directive": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Advance Directive"
          },
          "critical_information": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Critical Information"
          },
          "last_reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Reviewed At"
          },
          "organ_donor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OrganDonorStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "show_allergies": {
            "default": true,
            "title": "Show Allergies",
            "type": "boolean"
          },
          "show_blood_group": {
            "default": true,
            "title": "Show Blood Group",
            "type": "boolean"
          },
          "show_conditions": {
            "default": false,
            "title": "Show Conditions",
            "type": "boolean"
          },
          "show_medications": {
            "default": false,
            "title": "Show Medications",
            "type": "boolean"
          }
        },
        "title": "EmergencyProfileOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OrganDonorStatus": {
        "enum": [
          "yes",
          "no",
          "undecided"
        ],
        "title": "OrganDonorStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Access token from /api/v1/auth/login",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Health Io API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/patients/{patient_id}/emergency-contacts": {
      "get": {
        "operationId": "list_contacts_api_v1_patients__patient_id__emergency_contacts_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ContactOut"
                  },
                  "title": "Response List Contacts Api V1 Patients  Patient Id  Emergency Contacts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Contacts",
        "tags": [
          "emergency"
        ]
      },
      "post": {
        "operationId": "add_contact_api_v1_patients__patient_id__emergency_contacts_post",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Contact",
        "tags": [
          "emergency"
        ]
      }
    },
    "/api/v1/patients/{patient_id}/emergency-contacts/{contact_id}": {
      "delete": {
        "operationId": "remove_contact_api_v1_patients__patient_id__emergency_contacts__contact_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "contact_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Contact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Remove Contact",
        "tags": [
          "emergency"
        ]
      }
    },
    "/api/v1/patients/{patient_id}/emergency-profile": {
      "get": {
        "operationId": "get_profile_api_v1_patients__patient_id__emergency_profile_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyProfileOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Profile",
        "tags": [
          "emergency"
        ]
      },
      "put": {
        "operationId": "save_profile_api_v1_patients__patient_id__emergency_profile_put",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmergencyProfileModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyProfileOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Save Profile",
        "tags": [
          "emergency"
        ]
      }
    },
...
  }
}